+++ lib/ex_mcp/acp/adapter_bridge/port_runner.ex
@@ -7,32 +7,41 @@ defmodule ExMCP.ACP.AdapterBridge.PortRunner do
     GOOGLE_API_KEY PI_API_KEY MIX_ENV MIX_TARGET
   )
 
+  @isolated_env_allowlist ~w(
+    HOME LANG LOGNAME NIX_SSL_CERT_FILE PATH SHELL SSL_CERT_DIR SSL_CERT_FILE
+    TEMP TMP TMPDIR TZ USER
+  )
+
   @spec open(String.t(), [String.t()], keyword(), module()) ::
           {:ok, port()} | {:error, term()}
   def open(cmd, args, opts, adapter_mod) do
-    executable = System.find_executable(cmd)
-
-    if executable do
-      cwd = Keyword.get(opts, :cwd, File.cwd!())
-
-      port_opts = [
-        :binary,
-        :exit_status,
-        :use_stdio,
-        :stderr_to_stdout,
-        args: Enum.map(args, &to_charlist/1),
-        cd: to_charlist(cwd),
-        env: safe_env(opts, adapter_mod)
-      ]
-
-      try do
-        port = Port.open({:spawn_executable, to_charlist(executable)}, port_opts)
-        {:ok, port}
-      catch
-        :error, reason -> {:error, {:port_open_failed, reason}}
-      end
+    with :ok <- validate_environment_policy(opts),
+         executable when is_binary(executable) <- System.find_executable(cmd) do
+      do_open(executable, args, opts, adapter_mod)
     else
-      {:error, {:executable_not_found, cmd}}
+      nil -> {:error, {:executable_not_found, cmd}}
+      {:error, _reason} = error -> error
+    end
+  end
+
+  defp do_open(executable, args, opts, adapter_mod) do
+    cwd = Keyword.get(opts, :cwd, File.cwd!())
+
+    port_opts = [
+      :binary,
+      :exit_status,
+      :use_stdio,
+      :stderr_to_stdout,
+      args: Enum.map(args, &to_charlist/1),
+      cd: to_charlist(cwd),
+      env: safe_env(opts, adapter_mod)
+    ]
+
+    try do
+      port = Port.open({:spawn_executable, to_charlist(executable)}, port_opts)
+      {:ok, port}
+    catch
+      :error, reason -> {:error, {:port_open_failed, reason}}
     end
   end
 
@@ -56,11 +65,35 @@ defmodule ExMCP.ACP.AdapterBridge.PortRunner do
 
   @spec safe_env(keyword(), module()) :: [{charlist(), charlist() | false}]
   def safe_env(opts, adapter_mod) do
-    @session_vars_to_clear
-    |> Map.new(&{&1, false})
+    opts
+    |> Keyword.get(:environment_policy, :isolated)
+    |> base_env()
+    |> Map.merge(Map.new(@session_vars_to_clear, &{&1, false}))
     |> Map.put("TERM", "dumb")
     |> Map.merge(adapter_env(opts, adapter_mod))
     |> port_env()
+  end
+
+  defp base_env(:inherit), do: %{}
+
+  defp base_env(:isolated) do
+    parent_env = System.get_env()
+
+    retained =
+      Map.filter(parent_env, fn {name, _value} ->
+        name in @isolated_env_allowlist or String.starts_with?(name, "LC_")
+      end)
+
+    parent_env
+    |> Map.new(fn {name, _value} -> {name, false} end)
+    |> Map.merge(retained)
+  end
+
+  defp validate_environment_policy(opts) do
+    case Keyword.get(opts, :environment_policy, :isolated) do
+      policy when policy in [:isolated, :inherit] -> :ok
+      policy -> {:error, {:invalid_environment_policy, policy}}
+    end
   end
 
   defp adapter_env(opts, adapter_mod) do